Limit test to macOS 15 which sidesteps an Xcode version with a static linker bug.
Add .p2align directives to the assembly.

Change the naming so we have $SRCDIR/interrupt-and-trap-funcs.s
which is preprocessed into $OBJDIR/interrupt-and-trap-funcs.s
and then compiled into interrupt-and-trap-funcs.o.  This may
help avoid the addition of C headers I was getting on one Linux
bot when I was pretending it was a .c file.